+++ libtls/build.rs
@@ -1,0 +1,5 @@
+fn main() {
+    if libtls_sys::TLS_API >= 20_20_01_20 {
+        println!("cargo:rustc-cfg=libressl_3_1_0");
+    }
+}